feat(multithread): Remove ErrorSummary. Use ErrorMetricsKind in ErrorMetricsParser
