+++ iai-callgrind-runner/src/runner/tool/logfile_parser.rs
@@ -1,21 +1,16 @@
-use std::fs::File;
-use std::io::{BufRead, BufReader};
 use std::path::PathBuf;
-use std::str::FromStr;
 
-use anyhow::{Context, Result};
+use anyhow::Result;
 use lazy_static::lazy_static;
 use log::debug;
 use regex::Regex;
 
+use super::error_metric_parser::ErrorMetricLogfileParser;
 use super::{ToolOutputPath, ValgrindTool};
-use crate::error::Error;
-use crate::runner::costs::Costs;
+use crate::api::ErrorMetricKind;
 use crate::runner::dhat::logfile_parser::DhatLogfileParser;
-use crate::runner::summary::{
-    CostsKind, CostsSummary, CostsSummaryType, ErrorSummary, ToolRunSummary,
-};
-use crate::util::{make_relative, EitherOrBoth};
+use crate::runner::summary::{CostsKind, CostsSummary, CostsSummaryType, ToolRunSummary};
+use crate::util::EitherOrBoth;
 
 // The different regex have to consider --time-stamp=yes
 lazy_static! {
@@ -47,16 +42,8 @@ pub struct LogfileSummary {
     pub parent_pid: Option<i32>,
     pub fields: Vec<(String, String)>,
     pub details: Vec<String>,
-    pub error_summary: Option<ErrorSummary>,
     pub log_path: PathBuf,
     pub costs: CostsKind,
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
-enum State {
-    Header,
-    HeaderSpace,
-    Body,
 }
 
 /// TODO: Use the Parser trait instead if possible
@@ -107,7 +94,6 @@ impl LogfileSummary {
             parent_pid: None,
             summary: self.fields.into_iter().collect(),
             details: (!self.details.is_empty()).then(|| self.details.join("\n")),
-            error_summary: self.error_summary,
             log_path: self.log_path,
             costs_summary: CostsSummaryType::default(),
         }
@@ -184,19 +170,6 @@ impl LogfileSummary {
             _ => panic!("The logfile summaries of new and old costs should match"),
         };
 
-        //     (None, None) => None,
-        //     (None, Some(old_costs)) => {
-        //         Some(CostsSummary::new(EitherOrBoth::Right(old_costs.clone())))
-        //     }
-        //     (Some(new_costs), None) => {
-        //         Some(CostsSummary::new(EitherOrBoth::Left(new_costs.clone())))
-        //     }
-        //     (Some(new_costs), Some(old_costs)) => Some(CostsSummary::new(EitherOrBoth::Both((
-        //         new_costs.clone(),
-        //         old_costs.clone(),
-        //     )))),
-        // };
-
         let old_pid = Some(old.pid);
         let old_parent_pid = old.parent_pid;
         let pid = Some(self.pid);
@@ -209,111 +182,6 @@ impl LogfileSummary {
             costs_summary,
             ..self.raw_into_tool_run()
         }
-    }
-}
-
-struct ErrorMetricLogfileParser {
-    root_dir: PathBuf,
-}
-
-impl LogfileParser for ErrorMetricLogfileParser {
-    fn parse_single(&self, path: PathBuf) -> Result<LogfileSummary> {
-        let file = File::open(&path)
-            .with_context(|| format!("Error opening log file '{}'", path.display()))?;
-
-        let mut iter = BufReader::new(file)
-            .lines()
-            .map(std::result::Result::unwrap)
-            .skip_while(|l| l.trim().is_empty());
-
-        let line = iter
-            .next()
-            .ok_or_else(|| Error::ParseError((path.clone(), "Empty file".to_owned())))?;
-        let pid = extract_pid(&line);
-
-        let mut state = State::Header;
-        let mut command = None;
-        let mut details = vec![];
-        let mut error_summary = None;
-        let mut parent_pid = None;
-        for line in iter {
-            match &state {
-                State::Header if !EMPTY_LINE_RE.is_match(&line) => {
-                    if let Some(caps) = EXTRACT_FIELDS_RE.captures(&line) {
-                        let key = caps.name("key").unwrap().as_str();
-                        match key.to_ascii_lowercase().as_str() {
-                            "command" => {
-                                let value = caps.name("value").unwrap().as_str();
-                                command = Some(make_relative(&self.root_dir, value));
-                            }
-                            "parent pid" => {
-                                let value = caps.name("value").unwrap().as_str().to_owned();
-                                parent_pid = Some(
-                                    value
-                                        .as_str()
-                                        .parse::<i32>()
-                                        .expect("Parent PID should be valid"),
-                                );
-                            }
-                            _ => {}
-                        }
-                    }
-                }
-                State::Header => state = State::HeaderSpace,
-                State::HeaderSpace if EMPTY_LINE_RE.is_match(&line) => {}
-                State::HeaderSpace | State::Body => {
-                    if state == State::HeaderSpace {
-                        state = State::Body;
-                    }
-                    if let Some(caps) = EXTRACT_FIELDS_RE.captures(&line) {
-                        let key = caps.name("key").unwrap().as_str();
-                        if key.eq_ignore_ascii_case("error summary") {
-                            let error_summary_value = caps.name("value").unwrap().as_str();
-                            error_summary =
-                                ErrorSummary::from_str(error_summary_value).map(Some)?;
-
-                            continue;
-                        }
-                    }
-                    if let Some(caps) = STRIP_PREFIX_RE.captures(&line) {
-                        let rest_of_line = caps.name("rest").unwrap().as_str();
-                        details.push(rest_of_line.to_owned());
-                    } else {
-                        details.push(line);
-                    }
-                }
-            }
-        }
-
-        while let Some(last) = details.last() {
-            if last.trim().is_empty() {
-                details.pop();
-            } else {
-                break;
-            }
-        }
-
-        Ok(LogfileSummary {
-            command: command.expect("A command should be present"),
-            pid,
-            parent_pid,
-            fields: Vec::default(),
-            details,
-            error_summary,
-            log_path: make_relative(&self.root_dir, path),
-            // TODO: FIX
-            costs: CostsKind::ErrorCosts(Costs::empty()),
-        })
-    }
-
-    fn merge_logfile_summaries(
-        &self,
-        _: Vec<LogfileSummary>,
-        new: Vec<LogfileSummary>,
-    ) -> Vec<ToolRunSummary> {
-        new.into_iter()
-            .map(LogfileSummary::new_into_tool_run)
-            .collect()
     }
 }
 
@@ -339,8 +207,11 @@ impl ValgrindTool {
 
 impl LogfileSummary {
     pub fn has_errors(&self) -> bool {
-        self.error_summary
-            .as_ref()
-            .map_or(false, ErrorSummary::has_errors)
+        match &self.costs {
+            CostsKind::None | CostsKind::DhatCosts(_) => false,
+            CostsKind::ErrorCosts(costs) => costs
+                .cost_by_kind(&ErrorMetricKind::Errors)
+                .map_or(false, |e| e > 0),
+        }
     }
 }
